Game based on Bonus-Features2 by "Unity"
The project is borrowed from Unity Junior Programmer course. Aim is to update this game with new features. The original of project package file is in project folder.
Initial commit is made after adapting PlayerController.cs script to two characters and original scene is split in two sections for two players.
In "Starter" folder appears to be the original version of project.
In "Final" folder is all my updated version of project.